Extract _pid.py from _legacy.py
